Change export folder structure and chunk filenames
This commit changes the way how files are exported so that they match
with how other existing tools do it.

The file structure will now end up like this:
 - <ROOT>/bundles/ebx/<EBX>
 - <ROOT>/bundles/res/<RESOURCE>
 - <ROOT>/bundles/chunks/<CHUNK>
 - <ROOT>/chunks/<TOC RESOURCE>

The parsing logic for Toc resources/chunks is also modified to read a
GUID value instead of an SHA1 sum. To support this in the model, a new
class TocResource is added that acts as the base for both Toc resources
and bundle chunks (both use a GUID).

The SHA1 attribute is also made optional as it appears that Toc
resources do not have it.

Lastly, the filenames of chunks are now formatted as <GUID>.chunk.